+++ custom_components/sugar_valley_neopool/binary_sensor.py
@@ -70,7 +70,7 @@ BINARY_SENSOR_DESCRIPTIONS: tuple[NeoPoolBinarySensorEntityDescription, ...] = (
         key="modules_conductivity",
         translation_key="modules_conductivity",
         name="Conductivity Module",
-        icon="mdi:flash-circle",
+        icon="mdi:resistor",
         json_path="NeoPool.Modules.Conductivity",
     ),
     NeoPoolBinarySensorEntityDescription(